move autorandr into home-manager
While autorandr is technically something that should behave
differently depending on the system, enabling it system-wide
when xfce itself is enabled by home-manager is a bit yucky.
Also, the linear filtering option isn't present on the nixos
end, but it's present in home-manager, so that whole module
can be removed.

This change is subject to be reverted.